Add cache invalidation for mutations and fix column types to BIGINT
- Add cache invalidation calls to useVotePost, useEditPost, useDeletePost,
  useAddReply, and useCreateThread mutations
- Add DELETE /cache/threads and DELETE /cache/thread/:threadId endpoints
  to lambda-cache for invalidating public cache on content changes
- Fix forum_posts.likes, dislikes, reply_count to BIGINT in schema.sql
  and schema_migration.sql to match function return types

+++ schema.sql
@@ -237,9 +237,9 @@ CREATE TABLE IF NOT EXISTS forum_posts (
   created_at TIMESTAMPTZ DEFAULT NOW(),
   edited_at TIMESTAMPTZ,
   search_document TSVECTOR,
-  likes INTEGER DEFAULT 0,
-  dislikes INTEGER DEFAULT 0,
-  reply_count INTEGER DEFAULT 0
+  likes BIGINT DEFAULT 0,
+  dislikes BIGINT DEFAULT 0,
+  reply_count BIGINT DEFAULT 0
 );
 
 CREATE INDEX IF NOT EXISTS idx_forum_posts_thread ON forum_posts(thread_id);
